DSBD-003 대시보드 조회 API 작성

+++ src/main/kotlin/com/midasit/mcafe/domain/looseHistory/dto/LooseHistoryDto.kt
@@ -6,6 +6,6 @@ import com.midasit.mcafe.model.GameType
 data class LooseHistoryDto(
     val looser: Member,
     val gameType: GameType,
-    val looseCount: Int
+    val looseCount: Long
 ) {
 }
